requirements.txt
asgiref==3.8.1
Django==5.1.5
Faker==33.3.1
joblib==1.4.2
numpy==1.26.4
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
scikit-learn==1.6.1
scikit-surprise==1.1.4
scipy==1.15.1
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
typing_extensions==4.12.2
tzdata==2024.2
